reduce CircularProgressParts to Circle
